refactor(infra): fold diag.Logger into logwatcher and picturewatcher (#197)
Remove internal/infrastructure/diag; each consumer package owns its Logger type and Nop/Std helpers.
