+++ internal/infrastructure/logwatcher/event_handler.go
@@ -4,11 +4,7 @@ import (
 	"time"
 
 	"vrchat-tweaker/internal/domain/activity"
-	"vrchat-tweaker/internal/infrastructure/diag"
 )
-
-// Logger is a diagnostic logger for logwatcher (shared with picturewatcher via diag).
-type Logger = diag.Logger
 
 // EventHandler receives parsed events from the watcher.
 type EventHandler interface {
