+++ internal/infrastructure/logwatcher/log_test.go
@@ -1,4 +1,4 @@
-package diag
+package logwatcher
 
 import "testing"
