+++ internal/infrastructure/logwatcher/multi_output_log_watcher.go
@@ -5,8 +5,6 @@ import (
 	"os"
 	"sync/atomic"
 	"time"
-
-	"vrchat-tweaker/internal/infrastructure/diag"
 )
 
 const defaultActiveLogStallTimeout = 60 * time.Second
@@ -48,7 +46,7 @@ func NewMultiOutputLogWatcher(
 	logger Logger,
 ) *MultiOutputLogWatcher {
 	if logger == nil {
-		logger = diag.Nop
+		logger = Nop
 	}
 	return &MultiOutputLogWatcher{
 		watchDir:              watchDir,
