+++ internal/infrastructure/picturewatcher/watcher_test.go
@@ -11,8 +11,6 @@ import (
 	"time"
 
 	"github.com/fsnotify/fsnotify"
-
-	"vrchat-tweaker/internal/infrastructure/diag"
 )
 
 func TestIsImageFile(t *testing.T) {
@@ -75,8 +73,8 @@ func (l *testLogger) Printf(format string, v ...any) {
 	l.mu.Unlock()
 }
 
-func (l *testLogger) asLogger() diag.Logger {
-	return diag.Logger(l.Printf)
+func (l *testLogger) asLogger() Logger {
+	return Logger(l.Printf)
 }
 
 func TestStart_ingestsNewImage(t *testing.T) {
@@ -155,7 +153,7 @@ func TestHandleEvent_ignoresNonImageCreateOp(t *testing.T) {
 		ctx:      ctx,
 		fsw:      mustWatcher(t),
 		ingest:   func(context.Context, string) error { calls++; return nil },
-		log:      diag.Nop,
+		log:      Nop,
 		debounce: time.Hour,
 	}
 	t.Cleanup(func() { _ = r.fsw.Close() })
@@ -178,7 +176,7 @@ func TestHandleEvent_ignoresChmodOnly(t *testing.T) {
 		ctx:      context.Background(),
 		fsw:      mustWatcher(t),
 		ingest:   func(context.Context, string) error { calls++; return nil },
-		log:      diag.Nop,
+		log:      Nop,
 		debounce: time.Hour,
 	}
 	t.Cleanup(func() { _ = r.fsw.Close() })
@@ -200,7 +198,7 @@ func TestHandleEvent_ingestsOnRename(t *testing.T) {
 		ctx:      context.Background(),
 		fsw:      mustWatcher(t),
 		ingest:   func(_ context.Context, p string) error { done <- p; return nil },
-		log:      diag.Nop,
+		log:      Nop,
 		debounce: 20 * time.Millisecond,
 	}
 	t.Cleanup(func() { _ = r.fsw.Close() })
@@ -222,7 +220,7 @@ func TestHandleEvent_ignoresNonImage(t *testing.T) {
 		ctx:      ctx,
 		fsw:      mustWatcher(t),
 		ingest:   func(context.Context, string) error { calls++; return nil },
-		log:      diag.Nop,
+		log:      Nop,
 		debounce: time.Hour,
 	}
 	t.Cleanup(func() { _ = r.fsw.Close() })
@@ -246,7 +244,7 @@ func TestHandleEvent_watchesNewDirectory(t *testing.T) {
 		ctx:      ctx,
 		fsw:      mustWatcher(t),
 		ingest:   func(context.Context, string) error { return nil },
-		log:      diag.Nop,
+		log:      Nop,
 		debounce: time.Hour,
 	}
 	t.Cleanup(func() { _ = r.fsw.Close() })
@@ -283,7 +281,7 @@ func TestFlush_respectsCancelledContext(t *testing.T) {
 	r := &run{
 		ctx:    ctx,
 		ingest: func(context.Context, string) error { calls++; return nil },
-		log:    diag.Nop,
+		log:    Nop,
 	}
 	r.pending = map[string]struct{}{"/a.png": {}, "/b.png": {}}
 	r.flush()
@@ -314,7 +312,7 @@ func TestRun_scheduleIngest_coalescesSamePath(t *testing.T) {
 	r := &run{
 		ctx:      ctx,
 		ingest:   ing,
-		log:      diag.Nop,
+		log:      Nop,
 		debounce: 50 * time.Millisecond,
 	}
 	r.scheduleIngest("/tmp/a.png")
@@ -341,7 +339,7 @@ func TestRun_scheduleIngest_flushesMultiplePaths(t *testing.T) {
 	r := &run{
 		ctx:      ctx,
 		ingest:   ing,
-		log:      diag.Nop,
+		log:      Nop,
 		debounce: 50 * time.Millisecond,
 	}
 	r.scheduleIngest("/a.png")
@@ -361,7 +359,7 @@ func TestLoop_exitsWhenWatcherClosed(t *testing.T) {
 		ctx:      context.Background(),
 		fsw:      w,
 		ingest:   func(context.Context, string) error { return nil },
-		log:      diag.Nop,
+		log:      Nop,
 		debounce: time.Second,
 	}
 	done := make(chan struct{})
@@ -425,7 +423,7 @@ func TestHandleEvent_ingestsOnWrite(t *testing.T) {
 		ctx:      ctx,
 		fsw:      mustWatcher(t),
 		ingest:   func(_ context.Context, p string) error { done <- p; return nil },
-		log:      diag.Nop,
+		log:      Nop,
 		debounce: 20 * time.Millisecond,
 	}
 	t.Cleanup(func() { _ = r.fsw.Close() })
@@ -454,7 +452,7 @@ func TestRun_stopFlushTimer_clearsPending(t *testing.T) {
 	r := &run{
 		ctx:      ctx,
 		ingest:   ing,
-		log:      diag.Nop,
+		log:      Nop,
 		debounce: 200 * time.Millisecond,
 	}
 	r.scheduleIngest("/x.png")
